Fix bug with no label_smoothing training
In build_criterion() function, there is a bug with loss_weight = torch.ones(vocab_size)
change to loss_weight = torch.ones(vocab_size, dtype=torch.half)

+++ PyTorch/Translation/GNMT/train.py
@@ -330,7 +330,7 @@ def set_iter_size(train_iter_size, train_global_batch_size, train_batch_size):
 def build_criterion(vocab_size, padding_idx, smoothing):
     if smoothing == 0.:
         logging.info(f'Building CrossEntropyLoss')
-        loss_weight = torch.ones(vocab_size)
+        loss_weight = torch.ones(vocab_size, dtype=torch.half)
         loss_weight[padding_idx] = 0
         criterion = nn.CrossEntropyLoss(weight=loss_weight, size_average=False)
     else:
